+++ cardinal/server/types/provider.go
@@ -3,13 +3,13 @@ package types
 import (
 	"pkg.world.dev/world-engine/cardinal/gamestate"
 	"pkg.world.dev/world-engine/cardinal/receipt"
-	"pkg.world.dev/world-engine/cardinal/server"
+	"pkg.world.dev/world-engine/cardinal/server/validator"
 	"pkg.world.dev/world-engine/cardinal/types"
 	"pkg.world.dev/world-engine/sign"
 )
 
 type ProviderWorld interface {
-	server.SignerAddressProvider
+	validator.SignerAddressProvider
 	UseNonce(signerAddress string, nonce uint64) error
 	GetSignerForPersonaTag(personaTag string, tick uint64) (addr string, err error)
 	AddTransaction(id types.MessageID, v any, sig *sign.Transaction) (uint64, types.TxHash)
